refactor(sdk): eventemitter typescript errors fixes and cleanup
