+++ src/utils/emitter/_EmitterSubscription.ts
@@ -1,24 +1,11 @@
-/**
- * Copyright (c) Facebook, Inc. and its affiliates.
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
- *
- * @format
- * @flow
- */
-
-'use strict';
-
-import type EventEmitter from './EventEmitter';
-import EventSubscription from './_EventSubscription';
-import type EventSubscriptionVendor from './_EventSubscriptionVendor';
+import { EventSubscription } from './_EventSubscription';
+import type { EventEmitter } from './EventEmitter';
+import type { EventSubscriptionVendor } from './_EventSubscriptionVendor';
 
 /**
  * EmitterSubscription represents a subscription with listener and context data.
  */
-export default class EmitterSubscription extends EventSubscription {
-  // $FlowFixMe[value-as-type]
+export class EmitterSubscription extends EventSubscription {
   emitter: EventEmitter;
   listener: Function;
   context: Object | null | undefined;
@@ -34,7 +21,6 @@ export default class EmitterSubscription extends EventSubscription {
    *   listener
    */
   constructor(
-    // $FlowFixMe[value-as-type]
     emitter: EventEmitter,
     subscriber: EventSubscriptionVendor,
     listener: Function,
