+++ src/utils/index.ts
@@ -1,2 +1,1 @@
 export * from './emitter/EventEmitter';
-export { default as EventEmitter } from './emitter/EventEmitter';
